I wanted to ensure that this version actually runs start to finish. For this I had to revert back to my yml file, the previous preconditioner.py file and an import modification in multitracer.py to find the Observations class correctly. I also created a copy of the yml file for a local run on   my laptop as opposed to the normal run on the carbon portal. The test run for --optimize was successful.
I now want to move to python 3.10 so we can use the same conda environment as runflex given that we may call it on the fly and also because skuggfaxe was updated to a new version of Ubuntu 22.04 which sort of also is an incentive to move to Python 3.10. I will postpone merging the changes from the main branch until after  I can run LumiaDA from the flex conda environment with Python 3.10.